unified errors, added new endpoints, added swagger;

+++ src/controllers/users.ts
@@ -1,119 +1,31 @@
-import { Request, Response } from 'express';
-import jwt from 'jsonwebtoken';
-import bcryptjs from 'bcryptjs';
+import { Response } from 'express';
 
 import UserModel from '../models/user';
-import { UserDocument, UserRole } from '../types/user';
-import { JWT_SECRET } from '../config/config';
+import { UserRole } from '../types/user';
 import { handleErrors } from '../config/handle-error';
 import { ExpressRequest } from '../types/expressRequest.interface';
+import { normalizeUser } from '../config/normalize-user';
 
-const normalizeUser = (user: UserDocument) => {
-  const token = jwt.sign({ id: user.id, email: user.email }, JWT_SECRET);
-  return {
-    firstName: user.firstName,
-    lastName: user.lastName,
-    username: user.username,
-    email: user.email,
-    phoneNumber: user.phoneNumber,
-    role: user.role,
-    newsletterConsent: user.newsletterConsent,
-    id: user.id,
-    token
-  }
-}
-
-const register = async (req: Request, res: Response) => {
-  try {
-    if (req.body.password !== req.body.repeatedPassword) {
-      res.status(400).json({ error: 'Passwords do not match' });
-      return;
-    }
-    const existingUser = await UserModel.findOne({ email: req.body.email });
-    if (existingUser) {
-      res.status(400).json({ error: 'Email already in use' });
-      return;
-    }
-    const newUser = new UserModel(req.body);
-    const savedUser = await newUser.save();
-    res.status(201).json(normalizeUser(savedUser));
-  } catch (err) {
-    handleErrors(err, res)
-  }
-}
-
-const login = async (req: Request, res: Response) => {
-  try {
-    const user = await UserModel.findOne({ email: req.body.email }).select('+password');
-    if (!user) {
-      res.status(422).json({ error: 'Incorrect email' });
-      return;
-    }
-    const isSamePassword = await user.validatePassword(req.body.password);
-    if (!isSamePassword) {
-      res.status(401).json({ error: 'Incorrect password' });
-      return;
-    }
-    res.status(201).json(normalizeUser(user));
-  } catch (err) {
-    handleErrors(err, res)
-  }
-}
 
 const currentUser = async (req: ExpressRequest, res: Response) => {
   if (!req.user) {
-    res.status(401).json({ error: 'Unauthorized' });
+    res.status(401).json({
+      error: 'Unauthorized',
+      message: 'You are not authorized to access this resource.'
+    });
     return;
   }
   res.send(normalizeUser(req.user))
 }
-// todo need to test // todo add description to postman
-const checkUserByEmail = async (req: ExpressRequest, res: Response) => {
-  try {
-    const { email } = req.query;
-    if (!email) {
-      res.status(400).json({ error: 'Email is required' });
-      return;
-    }
-    const user = await UserModel.findOne({ email: String(email) });
-    if (!user) {
-      res.status(404).json({ error: 'User not found' });
-      return;
-    }
-    res.status(200).json({ message: 'User found' });
-  } catch (error) {
-    handleErrors(error, res)
-  }
-}
-// todo add description to postman
-const resetPassword = async (req: ExpressRequest, res: Response) => {
-  try {
-    const { email, newPassword } = req.body;
-    if (!email || !newPassword) {
-      res.status(400).json({ error: 'Email and new password are required' });
-      return;
-    }
-    const user = await UserModel.findOne({ email });
-    if (!user) {
-      res.status(404).json({ error: 'User not found' });
-      return;
-    }
-    const salt = await bcryptjs.genSalt(10);
-    const hashedPassword = await bcryptjs.hash(newPassword, salt);
-    await UserModel.updateOne(
-      { email },
-      { $set: { password:  hashedPassword } }
-    );
-    res.status(200).json({ message: 'Password has been reset successfully' });
-  } catch (error) {
-    handleErrors(error, res)
-  }
-}
-// todo add description to postman
+
 const updateUser = async (req: ExpressRequest, res: Response) => {
+  // todo add validation that user can update only his own data
   try {
     if (!req.user) {
-      res.status(401).json({ error: 'Unauthorized' });
+      res.status(401).json({
+        error: 'Unauthorized',
+        message: 'You are not authorized to access this resource.'
+      });
       return;
     }
     const updatedUser = await UserModel.findByIdAndUpdate(req.body.id, req.body, {new: true});
@@ -145,16 +57,22 @@ const getUsersList = async (req: ExpressRequest, res: Response) => {
     handleErrors(err, res)
   }
 }
-// todo add description to postman
+
 const getUserById = async (req: ExpressRequest, res: Response) => {
   try {
     const user = await UserModel.findById(req.params.userId);
     if (!user) {
-      res.status(404).json({ error: 'User not found' });
+      res.status(404).json({
+        error: 'User not found',
+        message: 'The user with the specified ID was not found.'
+      });
       return;
     }
     if (user.role === UserRole.ADMIN && req.user && user.id !== req.user.id) {
-      res.status(403).json({ error: 'Access denied' });
+      res.status(403).json({
+        error: 'Access denied',
+        message: 'You do not have permission to access this resource.'
+      });
       return;
     }
     res.status(200).json(normalizeUser(user));
@@ -162,22 +80,34 @@ const getUserById = async (req: ExpressRequest, res: Response) => {
     handleErrors(err, res)
   }
 }
-// todo need to test // todo add description to postman
+
+// todo need to test (added logic of deleting without id)
 const deleteUser = async (req: ExpressRequest, res: Response) => {
   try {
+    const user = req.user;
+    if (!user) {
+      res.status(404).json({
+        error: 'User not found',
+        message: 'The user with the specified ID was not found.'
+      });
+      return;
+    }
+    if (user.role === UserRole.ADMIN && req.user && user.id !== req.user.id) {
+      res.status(403).json({
+        error: 'Access denied',
+        message: 'You do not have permission to access this resource.'
+      });
+      return;
+    }
     await UserModel.deleteOne({ _id: req.params.userId });
-    res.status(200).json({ message: 'User deleted successfully' });
+    res.status(204).end();
   } catch (err) {
     handleErrors(err, res)
   }
 }
 
 export default {
-  register,
-  login,
-  currentUser,
-  checkUserByEmail,
-  resetPassword,
+  // currentUser,
   updateUser,
   getUsersList,
   getUserById,
